app/requirements.txt
aniso8601==9.0.1
attrs==23.2.0
audioread==3.0.1
azure-cognitiveservices-speech==1.36.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
dtw==1.4.0
fastdtw==0.3.4
Flask==3.0.2
flask-restx==1.3.0
fonttools==4.51.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-texttospeech==2.16.3
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
importlib_resources==6.4.0
itsdangerous==2.1.2
jaconv==0.3.4
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mecab-python3==1.0.8
msgpack==1.0.8
numba==0.59.1
numpy==1.26.4
packaging==24.0
pillow==10.3.0
platformdirs==4.2.0
pooch==1.8.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.13.0
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
style==1.1.0
threadpoolctl==3.4.0
typing_extensions==4.10.0
update==0.0.1
urllib3==2.2.1
Werkzeug==3.0.1
zhon==2.0.2
